codesense_backend_central/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
pymongo
python-dotenv
bcrypt
cryptography
PyJWT